requirements.txt
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0
google-api-python-client>=2.0.0
mcp>=1.0.0
